+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 2.0.0 (2020-09-09)
+* Adds PAdES Signer
+* Adds support for Trust Services Management
+
 ## 1.0.0 (2018-04-11)
 * First publicly available version